adding files

+++ JourneyPlanning.cpp
@@ -10,16 +10,16 @@ using namespace std;
 
 int main()
 {
-    int n;cin>>n;
-    vector<int> a(n);
-    map<int,vector<int> >m;
-    for(int i=0;i<n;i++){
+    long long n;cin>>n;
+    vector<long long> a(n);
+    map<long long,vector<long long> >m;
+    for(long long i=0;i<n;i++){
         cin>>a[i];
         m[a[i]-i].push_back(i);
     }   
-    int ans=0;
+    long long ans=0;
     for(auto i:m){
-        int sum=0;
+        long long sum=0;
         for(auto j:i.second){
             sum+=a[j];
         }
